+++ src/app/signup/page.tsx
@@ -4,7 +4,6 @@ import Signup from "@/widgets/account/ui/Signup";
 import SignupProfilePage from "@/app/signup/profile/page";
 
 export default function SignupPage() {
-  
   return (
     <SignupProfilePage/>
   )
