+++ src/entities/signup/ui/SelectBirthday.tsx
@@ -7,33 +7,17 @@ import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
 import { CalendarIcon } from 'lucide-react'
 import { cn } from '@/lib/utils'
-
-// Jotai atom for managing birthday state
-export const birthdayAtom = atom<string>('')
+import {birthdayAtom} from "@/store/signupStore";
+import useSelectBirthday from "@/entities/signup/hooks/useSelectBirthday";
 
 export default function SelectBirthday() {
-  const [birthday, setBirthday] = useAtom(birthdayAtom)
-  const [isFocused, setIsFocused] = useState(false)
-  
-  const handleBirthdayChange = (value: string) => {
-    // 숫자만 입력 허용
-    const numbersOnly = value.replace(/[^\d]/g, '')
-    
-    // 최대 8자리까지만 허용 (YYYYMMDD)
-    if (numbersOnly.length <= 8) {
-      setBirthday(numbersOnly)
-    }
-  }
-  
-  // 생년월일 포맷팅 (YYYY-MM-DD 형태로 표시)
-  const formatBirthday = (value: string) => {
-    if (value.length <= 4) return value
-    if (value.length <= 6) return `${value.slice(0, 4)}-${value.slice(4)}`
-    return `${value.slice(0, 4)}-${value.slice(4, 6)}-${value.slice(6, 8)}`
-  }
-  
-  const displayValue = formatBirthday(birthday)
-  const isValid = birthday.length === 8
+  const {
+    isFocused,
+    setIsFocused,
+    handleBirthdayChange,
+    displayValue,
+    isValid,
+  } = useSelectBirthday();
   
   return (
     <div className="space-y-4 mt-12">
