+++ src/entities/signup/ui/SetNickname.tsx
@@ -7,88 +7,16 @@ import { CheckCircle2, XCircle, Loader2 } from 'lucide-react'
 import { cn } from '@/lib/utils'
 
 import { validateNickname, checkNicknameDuplicate } from '@/utils/nicknameUtils'
-import { useNickname, useNicknameActions } from "@/entities/signup/hooks/useSetNickname"
 import { NicknameInputStatus } from "@/entities/signup/model/types"
+import useNickname from "@/entities/signup/hooks/useNickname";
 
 export default function SetNickname() {
-  const { nickname, validation } = useNickname()
-  const { updateNickname, updateValidation } = useNicknameActions()
-  const [debounceTimer, setDebounceTimer] = useState<number | null>(null)
-  
-  /**
-   * 닉네임 입력 핸들러
-   */
-  const handleNicknameChange = (value: string) => {
-    updateNickname(value)
-    
-    // 기본 유효성 검사
-    const basicValidation = validateNickname(value)
-    
-    if (!basicValidation.isValid) {
-      updateValidation({
-        isValid: false,
-        isDuplicate: false,
-        isChecking: false,
-        errorMessage: basicValidation.errorMessage
-      })
-      return
-    }
-    
-    // 디바운스로 중복 체크
-    if (debounceTimer) {
-      clearTimeout(debounceTimer)
-    }
-    
-    updateValidation({
-      isChecking: true,
-      errorMessage: ''
-    })
-    
-    const timer = setTimeout(() => {
-      checkNicknameDuplicate(value)
-        .then((isDuplicate) => {
-          updateValidation({
-            isValid: !isDuplicate,
-            isDuplicate,
-            isChecking: false,
-            errorMessage: isDuplicate ? '중복된 별명입니다.' : ''
-          })
-        })
-        .catch(() => {
-          updateValidation({
-            isValid: false,
-            isDuplicate: false,
-            isChecking: false,
-            errorMessage: '닉네임 확인 중 오류가 발생했습니다.'
-          })
-        })
-    }, 500)
-    
-    setDebounceTimer(timer)
-  }
-  
-  /**
-   * 입력 상태 계산
-   */
-  const getInputStatus = (): NicknameInputStatus => {
-    if (validation.isChecking) return 'checking'
-    if (validation.isDuplicate) return 'duplicate'
-    if (validation.isValid && nickname.trim()) return 'valid'
-    if (validation.errorMessage) return 'error'
-    return 'default'
-  }
-  
-  // 컴포넌트 언마운트 시 타이머 정리
-  useEffect(() => {
-    return () => {
-      if (debounceTimer) {
-        clearTimeout(debounceTimer)
-      }
-    }
-  }, [debounceTimer])
-  
-  const inputStatus = getInputStatus()
-  
+  const {
+    nickname,
+    validation,
+    inputStatus,
+    handleNicknameChange,
+  } = useNickname();
   return (
     <div className="space-y-6">
       {/* 닉네임 입력 필드 */}
